Add transactions to backup file - fix bug with useMidenClient hook that caused the miden client to repeatedly reinitialize

+++ src/lib/miden/reset.ts
@@ -2,9 +2,11 @@ import browser from 'webextension-polyfill';
 
 import * as Repo from 'lib/miden/repo';
 
-export async function clearStorage() {
-  await Repo.db.delete();
-  await Repo.db.open();
+export async function clearStorage(clearDb: boolean = true) {
+  if (clearDb) {
+    await Repo.db.delete();
+    await Repo.db.open();
+  }
   await browser.storage.local.clear();
 }
